pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "subduedwealthypup"
version = "0.1.0"
[tool.poetry.dependencies]
flask = "^1.1"
python = "^3.7"
dnspython = "^1.16"
pymongo = "^3.10"
[tool.poetry.dev-dependencies]
